event venue destroy method now check if venue contains events

+++ src/Http/Controllers/EventVenueController.php
@@ -190,10 +190,15 @@ class EventVenueController extends Controller
      */
     public function destroy(EventVenue $eventVenue)
     {
-        $eventVenue->delete();
-
-        return redirect()->route('eventVenues.index')
-                        ->with('success', __('laravel-events-calendar::messages.venue_deleted_successfully'));
+        if (EventVenue::venueContainsEvents($eventVenue->id)){
+            return redirect()->route('eventVenues.index')
+                            ->with('success', __('laravel-events-calendar::messages.venue_not_deleted'));
+        }
+        else{
+            $eventVenue->delete();
+            return redirect()->route('eventVenues.index')
+                            ->with('success', __('laravel-events-calendar::messages.venue_deleted_successfully'));
+        }
     }
 
     /***************************************************************************/
